move morphism into submodule
